// frontend/types.ts
export enum QuestionType {
  SINGLE_CHOICE = "single_choice",
  MULTIPLE_CHOICE = "multiple_choice",
  TRUE_FALSE = "true_false",
}

export interface Question {
  id: string;
  text: string;
  type: QuestionType;
  options: string[];
  correctAnswers: number[]; // Index of options
  moduleName?: string;
}

export interface Test {
  id: string;
  name: string;
  duration: number; // Mins
  passingMarks: number;
  questions: Question[];
  isPublished: boolean;
  createdAt: string;
}

export enum TestStatus {
  NOT_STARTED = "not_started",
  IN_PROGRESS = "in_progress",
  COMPLETED = "completed",
}

export interface AssignedTest {
  id: string; // unique assigned test id
  testId: string;
  testName: string;
  employeeId: string;
  status: TestStatus;
  score?: number;
  totalQuestions?: number;
  passingMarks?: number;
  passed?: boolean;
  answers?: Record<string, number[]>; // questionId -> selectedOptionIndices
  remainingTime?: number; // local state tracker, in seconds
  startedAt?: string;
  completedAt?: string;
  assignedAt?: string;
  isDefaultOnboardingTest?: boolean;
}